Remove unused HSA stuff
