+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.4.15"
 [deps]
 AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 BinaryProvider = "b99e7846-7c00-51b0-8f62-c81ae34c0232"
 CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
 ExprTools = "e2ba6199-217a-4e67-a87a-7c52f15ade04"
@@ -20,14 +19,11 @@ LLVM_jll = "86de99a1-58d6-5da7-8064-bd56ce2e322c"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
-MsgPack = "99f44e22-a591-53d1-9472-aa23ef4bd671"
-ObjectFile = "d8793406-e978-5875-9003-1fc021f44a92"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ROCmDeviceLibs_jll = "873c0968-716b-5aa7-bb8d-d1e2e2aeff2d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Setfield = "efcf1570-3423-57d1-acb7-fd33fddbac46"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TimespanLogging = "a526e669-04d3-4846-9525-c66122c55f63"
@@ -51,11 +47,8 @@ LLD_jll = "12, 13, 14, 15"
 LLVM = "5"
 LLVM_jll = "12, 13, 14, 15"
 MacroTools = "0.5"
-MsgPack = "1"
-ObjectFile = "0.3"
 Preferences = "1"
 ROCmDeviceLibs_jll = "4, 5"
-Setfield = "0.5, 0.6, 0.7, 1.0"
 SpecialFunctions = "1, 2"
 TimespanLogging = "0.1"
 UnsafeAtomicsLLVM = "0.1"
